+++ R/week_chart.R
@@ -25,7 +25,7 @@ NULL
 #' @export
 week_chart <- function(wvalue, lgnm = "Value",
                        high = "yellow", low = "green", width = 0.9){
-  stopifnot("Give a numeric vector" = is.numeric(x))
+  stopifnot("Give a numeric vector" = is.numeric(wvalue))
   days <- paste(c("Satur", "Sun", "Mon", "Tues", "Wednes", "Thurs", "Fri"), "day", sep = "")
   df <- tibble::tibble(days, wvalue) %>%
     dplyr::mutate(days = factor(days, levels = days))